Support Django 1.9. Fixes #7. Fixes #9
