+++ docs/installation.txt
@@ -26,14 +26,13 @@ Installation
     from django_autoconfig.autoconfig import configure_settings
     configure_settings(globals())
 
-* For production, add the following settings for django-compressor_ (see their documentation for more detail)::
+* For production, add the following settings for django-pipeline_ (see their documentation for more detail)::
 
-    COMPRESS_OFFLINE = True
+    PIPELINE_ENABLED = True
 
-* Run ``collectstatic`` and ``compress`` (``compress`` is only needed if ``COMPRESS_OFFLINE`` is ``True``)::
+* Run ``collectstatic``::
 
     manage.py collectstatic
-    manage.py compress
 
 * You're now ready to start using Nuit in your templates and apps.
 
@@ -48,11 +47,11 @@ In case you're installing Nuit differently (e.g. from the git repo), make sure t
 
       pip install django
 
-- django-compressor_
+- django-pipeline_
 
   In order to compile the CSS used by Nuit::
 
-      pip install django-compressor
+      pip install django-pipeline
 
 - django-foundation-statics_
 
@@ -79,7 +78,7 @@ In case you're installing Nuit differently (e.g. from the git repo), make sure t
       pip install django-autoconfig
 
 .. _SASS: http://sass-lang.com
-.. _`django-compressor`: http://django-compressor.readthedocs.org/en/latest/
+.. _`django-pipeline`: https://django-pipeline.readthedocs.org/en/latest/
 .. _`django`: https://www.djangoproject.com/
 .. _`django-foundation-statics`: https://pypi.python.org/pypi/django-foundation-statics/
 .. _`Foundation`: http://foundation.zurb.com
@@ -87,4 +86,3 @@ In case you're installing Nuit differently (e.g. from the git repo), make sure t
 .. _`Foundation Icons`: http://zurb.com/playground/foundation-icon-fonts-3
 .. _`django-bourbon`: https://pypi.python.org/pypi/django-bourbon/
 .. _`django-autoconfig`: https://pypi.python.org/pypi/django-autoconfig/
-
