+++ example_project/demo/templates/demo/base.html
@@ -1,3 +1,4 @@
+{% load nuit %}
 {% extend 'nuit/base.html' topbar=True leftmenu=True breadcrumbs=True show_user_info=True show_search=True search_view='kitchen_sink' %}
 
 {% block breadcrumbs %}
