+++ example_project/demo/templates/demo/forms.html
@@ -1,3 +1,4 @@
+{% load nuit %}
 {% extend 'demo/base.html' rightmenu=True %}
 
 {% block title %}Forms{% endblock %}
