+++ example_project/demo/templates/demo/introduction.html
@@ -1,3 +1,4 @@
+{% load nuit %}
 {% extend 'demo/base.html' rightmenu=True app_logo='Nuit <span class="label radius">BETA</span>' %}
 
 {% block right_menu %}
@@ -67,11 +68,11 @@
                 <div>
                 <h5>Other Stuff</h5>
                 <div class="side-content">
-                
-                
+
+
         <p>This is some other stuff.</p>
-    
-                
+
+
                 </div>
                 </div>
                 <hr>
