+++ example_project/demo/templates/demo/kitchensink.html
@@ -1,3 +1,4 @@
+{% load nuit %}
 {% extend 'demo/base.html' rightmenu=True %}
 
 {% block title %}Kitchen Sink{% endblock %}
