+++ example_project/demo/templates/demo/list_view.html
@@ -1,3 +1,4 @@
+{% load nuit %}
 {% extend 'nuit/generic/list.html' base='demo/base.html' rightmenu=False topbar=True leftmenu=True rightmenu=True search_verb='Filter' breadcrumbs=True %}
 
 {% block title %}Publishers{% endblock %}
