+++ example_project/demo/templates/demo/table.html
@@ -1,3 +1,4 @@
+{% load nuit %}
 {% extend 'demo/base.html'  %}
 
 {% block content %}
