+++ nuit/templates/nuit/bases/_error_page.html
@@ -1,3 +1,4 @@
+{% load nuit %}
 {% extend nuit_base|default:'nuit/base.html' topbar=True topbar_classes='show-for-small-only' %}
 {% load static %}
