+++ nuit/templates/nuit/generic/400.html
@@ -1,1 +1,2 @@
+{% load nuit %}
 {% extend 'nuit/bases/_error_page.html' error_code=400 error_title="Bad Request" error_message="Sorry, you sent an invalid request." %}
