+++ nuit/templates/nuit/generic/403.html
@@ -1,1 +1,2 @@
+{% load nuit %}
 {% extend 'nuit/bases/_error_page.html' error_code=403 error_title="Permission Denied" error_message="Sorry, you don't have access to this page." %}
