+++ nuit/templates/nuit/generic/404.html
@@ -1,1 +1,2 @@
+{% load nuit %}
 {% extend 'nuit/bases/_error_page.html' error_code=404 error_title="Not Found" error_message="Sorry, we can't seem to find the page that you're after - it's likely you've followed a bad link." show_return=True %}
