+++ nuit/templates/nuit/generic/500.html
@@ -1,1 +1,2 @@
+{% load nuit %}
 {% extend 'nuit/bases/_error_page.html' error_code=500 error_title="Server Error" error_message="Sorry, it appears we've made a mistake and something has gone wrong. We'll look into it." %}
