+++ nuit/templates/nuit/generic/list.html
@@ -1,3 +1,4 @@
+{% load nuit %}
 {% extend nuit_base|default:'nuit/base.html' %}
 
 {% block content %}
