+++ setup.py
@@ -13,12 +13,12 @@ setup(
     packages             = find_packages(),
     install_requires     = [
         'django-foundation-statics >= 5.4.7-2',
-        'django_compressor >= 1.3.1ocado1, < 2.0',
         'django-jquery',
         'django-foundation-icons>=3.1',
-        'django < 1.8',
+        'django < 1.10',
         'django-autoconfig',
         'django-bourbon',
+        'django-pipeline >= 1.6.0',
     ],
     include_package_data = True,
     test_suite           = 'setuptest.setuptest.SetupTestSuite',
